siguen sin correr los tests

+++ Kernel/include/scheduler.h
@@ -12,7 +12,7 @@
 
 typedef struct schedulerCDT {
   process_t * processes[MAX_PROCESSES + 1];
-  int16_t current;
+  int64_t current;
   uint8_t size;
   uint16_t total_ticks;
 } schedulerCDT;
@@ -25,19 +25,19 @@ uint16_t total_ticks();
 void * schedule(void * context);
 int64_t add_process(entry_point_t main, char ** argv, char * name, uint8_t no_kill, int * file_descriptors);
 void destroy_scheduler();
-uint16_t get_current_pid();
+int64_t get_current_pid();
 void yield();
-int kill_process(uint16_t pid);
+int kill_process(int64_t pid);
 int32_t kill_current_process();
-int block_process(uint16_t pid);
-int unblock_process(uint16_t pid);
+int block_process(int64_t pid);
+int unblock_process(int64_t pid);
 int block_current_process();
 int unblock_current_process();
-int set_process_priority(uint16_t pid, uint8_t priority);
-int64_t wait_pid(uint16_t pid, int32_t * exit_code);
-int sleep_block(uint16_t pid, uint8_t sleep);
+int set_process_priority(int64_t pid, uint8_t priority);
+int64_t wait_pid(int64_t pid, int32_t * exit_code);
+int sleep_block(int64_t pid, uint8_t sleep);
 void my_exit(int64_t ret);
-uint8_t foreground_process(uint16_t pid);
+uint8_t foreground_process(int64_t pid);
 process_info_t * get_processes_info();
 
 #endif // SCHEDULER_H
