+++ Userland/SampleCodeModule/include/stdlib.h
@@ -39,22 +39,22 @@ void * my_malloc(uint64_t size);
 
 void my_free(void * ptr);
 
-uint64_t my_create_process(uint64_t main, char ** argv, char * name, uint8_t no_kill, int * file_descriptors);
+int64_t my_create_process(uint64_t main, char ** argv, char * name, uint8_t no_kill, int * file_descriptors);
 
-uint64_t my_wait_pid(int pid, int * exit_code);
+int64_t my_wait_pid(int64_t pid, int * exit_code);
 
 uint64_t my_exit(int exit_code);
 
 uint64_t my_yield();
 
-uint64_t my_getpid();
+int64_t my_getpid();
 
-uint64_t my_block_process(int16_t pid);
+int64_t my_block_process(int64_t pid);
 
-uint64_t my_unblock_process(int16_t pid);
+int64_t my_unblock_process(int64_t pid);
 
-uint64_t my_nice(int16_t pid, uint8_t priority);
+int64_t my_nice(int64_t pid, uint8_t priority);
 
-uint64_t my_kill(int16_t pid);
+int64_t my_kill(int64_t pid);
 
 #endif
