+++ Userland/SampleCodeModule/tests/test_util.c
@@ -70,7 +70,7 @@ void endless_loop_print(uint64_t wait) {
   int64_t pid = my_getpid();
 
   while (1) {
-    printf("%d ", pid);
+    printf("%lld ", (long long)pid);
     bussy_wait(wait);
   }
 }
